+++ types/peer.go
@@ -10,19 +10,20 @@ import (
 type P2P interface {
 	SendToValidators(msg proto.Message) ErrorI
 	SendTo(pubKey []byte, topic Topic, msg proto.Message) ErrorI
+	SendToAll(topic Topic, msg proto.Message) ErrorI
 	SendToPeer(topic Topic, msg proto.Message) (*PeerInfo, ErrorI)
 	ReceiveChannel(topic Topic) chan *MessageWrapper
-	GetPeerInfo(pubKey []byte) (PeerInfo, ErrorI)
-	GetPeerWithHeight(height uint64) *PeerInfo
-	GetMaxPeerHeight() uint64
+	GetPeerInfo(pubKey []byte) (*PeerInfo, ErrorI)
 	ChangeReputation(pubKey []byte, delta int32)
+	MaxPossiblePeers() int
 }
 
 type Channels map[Topic]chan *MessageWrapper
 
 type MessageWrapper struct {
 	Message proto.Message
-	Sender  PeerInfo
+	Hash    []byte
+	Sender  *PeerInfo
 }
 
 const (
@@ -30,10 +31,14 @@ const (
 )
 
 func (x *PeerInfo) AddressString() string {
+	return x.Address.AddressString()
+}
+
+func (x *PeerAddress) AddressString() string {
 	return BytesToString(x.PublicKey) + delimiter + x.NetAddress
 }
 
-func (x *PeerInfo) FromString(s string) ErrorI {
+func (x *PeerAddress) FromString(s string) ErrorI {
 	arr := strings.Split(s, delimiter)
 	if len(arr) != 2 {
 		return ErrInvalidNetAddrString(s)
@@ -52,4 +57,28 @@ func (x *PeerInfo) FromString(s string) ErrorI {
 	x.NetAddress = net.JoinHostPort(host, port)
 	x.PublicKey = pubKeyBytes
 	return nil
+}
+
+func (x *PeerAddress) Copy() *PeerAddress {
+	pkCopy := make([]byte, len(x.PublicKey))
+	copy(pkCopy, x.PublicKey)
+	return &PeerAddress{
+		PublicKey:  pkCopy,
+		NetAddress: x.NetAddress,
+	}
+}
+
+func (x *PeerInfo) Copy() *PeerInfo {
+	return &PeerInfo{
+		Address:     x.Address.Copy(),
+		IsOutbound:  x.IsOutbound,
+		IsValidator: x.IsValidator,
+		IsTrusted:   x.IsTrusted,
+		Reputation:  x.Reputation,
+	}
+}
+
+func (x *PeerInfo) FromString(s string) ErrorI {
+	x.Address = new(PeerAddress)
+	return x.Address.FromString(s)
 }